requirements.txt
arxiv
python-dotenv
anthropic
fastmcp
groq